+++ app/not-found.tsx
@@ -1,0 +1,12 @@
+export default function NotFound() {
+  return (
+    <div className="mt-8">
+      <div className="mx-auto max-w-4xl px-8">
+        <h1 className="text-4xl font-bold">404 - Not Found</h1>
+        <p className="mt-4">
+          The page you were looking for could not be found.
+        </p>
+      </div>
+    </div>
+  );
+}
