+++ components/SignInForm.tsx
@@ -1,5 +1,5 @@
 import { Dispatch, SetStateAction, useMemo, useState } from "react";
-import { useSession } from "../pages/_app";
+import { useSession } from "../components/SessionProvider";
 import { FormField } from "./FormField";
 import { FormErrorMessage } from "./FormErrorMessage";
