+++ package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "@headlessui/react": "^1.7.17",
     "@monaco-editor/react": "^4.6.0",
+    "@next/third-parties": "^14.1.4",
     "@prisma/client": "^4.1.0",
     "aws-sdk": "^2.1261.0",
     "bcrypt": "^5.1.1",
@@ -24,7 +25,7 @@
     "form-data": "^4.0.0",
     "jszip": "^3.10.1",
     "nanoid": "^5.0.4",
-    "next": "^14.1.0",
+    "next": "^14.1.4",
     "node-fetch": "^3.2.9",
     "raw-body": "^2.5.1",
     "react": "^18.2.0",
